+++ src/floating-select-windowed/index.js
@@ -1,0 +1,1 @@
+export FloatingSelectWindowed from './floating-select-windowed.component';